feat: UI를 위한 lottocontroller 생성
